backend/pyproject.toml
[tool.poetry]
name = "listic"
version = "0.1.0"
description = "AI-powered shopping list generator with store layout awareness"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "*"
djangorestframework = "*"
psycopg2-binary = "*"
python-dotenv = "*"
gunicorn = "*"
uvicorn = "*" # Added for ASGI support with Gunicorn

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
pytest = "^7.4.3"
pytest-django = "^4.5.2"
watchfiles = "*" # Add watchfiles for better reload support with Uvicorn

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
